// clients/typescript/src/types/common/messages.ts
export interface JsonRpcRequest {
  readonly jsonrpc: '2.0';
  readonly id: number;
  readonly method: string;
  readonly params?: unknown;
}

/** A JSON-RPC success response. */
export interface JsonRpcSuccessResponse {
  readonly jsonrpc: '2.0';
  readonly id: number;
  readonly result: unknown;
}
// ...
export interface JsonRpcNotification {
  readonly jsonrpc: '2.0';
  readonly method: string;
  readonly params?: unknown;
}